Revert "Kludge: work around lack of syscall.seek on 386 and arm, #1906"
This reverts commit 60b483bd3b63653ddff715bf2fa8a293ad998972.
