+++ src/os/stat_unix.go
@@ -15,6 +15,20 @@ func (f *File) Sync() error {
 	return ErrNotImplemented
 }
 
+// Stat returns the FileInfo structure describing file.
+// If there is an error, it will be of type *PathError.
+func (f *File) Stat() (FileInfo, error) {
+	var fs fileStat
+	err := ignoringEINTR(func() error {
+		return syscall.Fstat(int(f.handle.(unixFileHandle)), &fs.sys)
+	})
+	if err != nil {
+		return nil, &PathError{Op: "fstat", Path: f.name, Err: err}
+	}
+	fillFileStatFromSys(&fs, f.name)
+	return &fs, nil
+}
+
 // statNolog stats a file with no test logging.
 func statNolog(name string) (FileInfo, error) {
 	var fs fileStat
